bug fix: checkUnicFunc in the put methods + res.mw

+++ services/fighterService.js
@@ -1,3 +1,5 @@
+import HttpError from "../helpers/HttpError.js";
+import { checkFighterUnicName } from "../helpers/checkFighterUnicName.js";
 import { fighterRepository } from "../repositories/fighterRepository.js";
 
 class FighterService {
@@ -10,6 +12,9 @@ class FighterService {
   }
 
   create(credentials) {
+    const { name } = credentials;
+    const isUnicName = checkFighterUnicName(name);
+    if (!isUnicName) throw HttpError(409, "Conflict! Name is not available");
     const fighter = fighterRepository.create(credentials);
     return fighter;
   }
@@ -20,11 +25,16 @@ class FighterService {
   }
 
   patch(id, credentials) {
+    const { name } = credentials;
+    const isUnicName = checkFighterUnicName(name, id);
+    if (!isUnicName) throw HttpError(409, "Conflict! Name is not available");
     const fighter = fighterRepository.update(id, credentials);
     return fighter;
   }
 
   remove(id) {
+    const fighter = this.search({ id });
+    if (!fighter) throw HttpError(404, "Fighter not found");
     const response = fighterRepository.delete(id);
     return response;
   }
